Recommit all padocc scripts, removed Configuration in favour of GroupOperation and Mixins
