+++ padocc/__init__.py
@@ -1,1 +1,3 @@
-from .phases import Configuration
+__author__    = "Daniel Westwood"
+__contact__   = "daniel.westwood@stfc.ac.uk"
+__copyright__ = "Copyright 2024 United Kingdom Research and Innovation"
