+++ padocc/core/__init__.py
@@ -1,3 +1,7 @@
+__author__    = "Daniel Westwood"
+__contact__   = "daniel.westwood@stfc.ac.uk"
+__copyright__ = "Copyright 2024 United Kingdom Research and Innovation"
+
 from .logs import (
     init_logger, 
     reset_file_handler,
@@ -8,3 +12,5 @@ from .logs import (
 from .utils import (
     BypassSwitch
 )
+
+from .project import ProjectOperation
