+++ padocc/operations/__init__.py
@@ -1,8 +1,5 @@
-from .evaluate import EvaluateOperation
+__author__    = "Daniel Westwood"
+__contact__   = "daniel.westwood@stfc.ac.uk"
+__copyright__ = "Copyright 2024 United Kingdom Research and Innovation"
 
-from .operations import (
-    GroupOperation,
-    ProjectOperation,
-)
-
-from .allocate import create_allocations
+from .group import GroupOperation
